Mis listas se borra por ajax

+++ controllers/UsuariosListasController.php
@@ -3,8 +3,6 @@
 namespace app\controllers;
 
 use app\models\ListasProductos;
-use app\models\ListasSearch;
-use app\models\Usuarios;
 use Yii;
 use app\models\UsuariosListas;
 use app\models\UsuariosListasSearch;
@@ -163,12 +161,15 @@ class UsuariosListasController extends Controller
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
      */
-    public function actionDelete($id)
+    public function actionDelete()
     {
-        $model = $this->findModel($id);
-        $usuario_id = $model->usuario_id;
-        $model->delete();
-        return $this->redirect(['mis-listas', 'id' => $usuario_id]);
+        if (Yii::$app->request->isAjax) {
+            $id = Yii::$app->request->post('lista_id');
+            $model = $this->findModel($id);
+            $titulo = $model->lista->titulo;
+            $model->delete();
+            return $this->asJson(['titulo' => $titulo]);
+        }
     }
 
     /**
@@ -185,23 +186,5 @@ class UsuariosListasController extends Controller
         }
 
         throw new NotFoundHttpException('The requested page does not exist.');
-    }
-
-
-
-    /**
-     * Finds the Usuarios model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param integer $id
-     * @return Usuarios the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    protected function findUsuario($id)
-    {
-        if (($model = Usuarios::findOne($id)) !== null) {
-            return $model;
-        }
-
-        throw new NotFoundHttpException('La página no existe.');
     }
 }
